+++ sf_core/tests/integration/query/mod.rs
@@ -2,4 +2,5 @@ mod abort_query;
 mod json_result_set;
 mod json_result_set_large;
 mod json_result_set_nulls;
+mod operation_cancellation;
 mod statement_cancel;